move vqmodel to models.autoencoders.
